% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcop.R
\docType{data}
\name{anole_lizard}
\alias{anole_lizard}
\alias{c.elegans}
\alias{chicken}
\alias{chimpanzee}
\alias{cow}
\alias{dog}
\alias{fruitfly}
\alias{horse}
\alias{macaque}
\alias{mouse}
\alias{opossum}
\alias{pig}
\alias{platypus}
\alias{rat}
\alias{s.cerevisiae}
\alias{xenopus}
\alias{zebrafish}
\title{HCOP ortholog data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 88020 rows and 9 columns.}
\source{
\url{https://www.genenames.org/cgi-bin/hcop} via ftp at
  \url{ftp://ftp.ebi.ac.uk/pub/databases/genenames/hcop/}.
}
\usage{
anole_lizard

c.elegans

chicken

chimpanzee

cow

dog

fruitfly

horse

macaque

mouse

opossum

pig

platypus

rat

s.cerevisiae

xenopus

zebrafish
}
\description{
Human ortholog data from HGNC Comparison of Orthology Predictions (HCOP). The
HGNC Comparison of Orthology Predictions (HCOP) integrates and displays the
orthology assertions predicted for a specified human gene, or set of human
genes, by eggNOG, Ensembl Compara, HGNC, HomoloGene, Inparanoid, NCBI Gene
Orthology, OMA, OrthoDB, OrthoMCL, Panther, PhylomeDB, TreeFam and ZFIN. An
indication of the reliability of a prediction is provided by the number of
databases which concur. Data collected from HCOP at
<https://www.genenames.org/cgi-bin/hcop> via ftp at
<ftp://ftp.ebi.ac.uk/pub/databases/genenames/hcop/>.
}
\keyword{datasets}
