% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcop.R
\docType{data}
\name{chicken}
\alias{chicken}
\alias{fruitfly}
\alias{mouse}
\alias{rat}
\alias{xenopus}
\alias{zebrafish}
\title{HCOP ortholog data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46418 rows and 9 columns.}
\source{
\url{https://www.genenames.org/cgi-bin/hcop} via ftp at
  \url{ftp://ftp.ebi.ac.uk/pub/databases/genenames/hcop/}.
}
\usage{
chicken

fruitfly

mouse

rat

xenopus

zebrafish
}
\description{
Human ortholog data from HGNC Comparison of Orthology Predictions (HCOP). The
HGNC Comparison of Orthology Predictions (HCOP) integrates and displays the
orthology assertions predicted for a specified human gene, or set of human
genes, by eggNOG, Ensembl Compara, HGNC, HomoloGene, Inparanoid, NCBI Gene
Orthology, OMA, OrthoDB, OrthoMCL, Panther, PhylomeDB, TreeFam and ZFIN. An
indication of the reliability of a prediction is provided by the number of
databases which concur. This package only contains ortholog data to C.
elegans, chicken, fruitfly, mouse, rat, xenopus, and zebrafish. Others are
available at \url{https://www.genenames.org/cgi-bin/hcop} via ftp at
\url{ftp://ftp.ebi.ac.uk/pub/databases/genenames/hcop/}
}
\keyword{datasets}
